backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
openai>=1.42.0
openai-agents>=0.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
websockets==12.0
psutil==7.0.0
requests==2.31.0
typer==0.19.1
